ai-backend/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
anthropic==0.43.0
web3==7.6.0
python-dotenv==1.0.1
pydantic==2.10.4
httpx==0.28.1
